requirements.txt
tensorflow==2.13.0
torch==2.0.1
numpy==1.26.4
flask==2.3.2
scipy==1.14.0
scikit-learn==1.3.1
PyPDF2==3.0.1
pdfminer.six==20221105
pandas==2.1.2
nltk==3.8.1
gensim==4.3.2
spacy==3.7.5
transformers==4.42.3
onnx==1.14.0
statsmodels==0.14.0
prophet==1.2.4
matplotlib==3.7.2
easyocr==1.7.0
Pillow==10.0.1
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.1.0

